workaround C11 standards bug
This bug was fixed in C17, but we are using gcc 9.1, which caused
misalignment here on 32-bit platforms. We have this behind a lock
necessarily, so it could just be a boolean, but that might need some
more code cleanup.

+++ src/julia_threads.h
@@ -261,7 +261,7 @@ typedef struct _jl_tls_states_t {
     jl_gc_markqueue_t mark_queue;
     jl_gc_mark_cache_t gc_cache;
     arraylist_t sweep_objs;
-    _Atomic(int64_t) gc_sweeps_requested;
+    _Atomic(size_t) gc_sweeps_requested;
     // Saved exception for previous *external* API call or NULL if cleared.
     // Access via jl_exception_occurred().
     struct _jl_value_t *previous_exception;
